Add files via upload
This marksheet generator is a software tool or program designed to automate the process of creating academic or exam result sheets for students. It simplifies the task of recording and calculating marks, grades, and other relevant information, making it easier for educators and institutions to manage and distribute results to students and parents.